refactor: change structure of the project
